some pvm improve
